# Remove fmt call in the session for each call