Fixes #1632 

Running gateway headless with auth-key.

Before change:

```
Showing top 10 nodes out of 182
      flat  flat%   sum%        cum   cum%
     0.04s  0.58%  0.58%      0.04s  0.58%  fmt.(*fmt).fmt_sbx
     0.04s  0.58%  1.16%      0.06s  0.87%  github.com/TykTechnologies/tyk/vendor/github.com/TykTechnologies/redigocluster/rediscluster.ConcurrentMap.Iter.func1.1
     0.02s  0.29%  1.45%      0.04s  0.58%  net/textproto.CanonicalMIMEHeaderKey
     0.02s  0.29%  1.74%      0.02s  0.29%  net/textproto.validHeaderFieldByte (inline)
     0.02s  0.29%  2.03%      0.02s  0.29%  strings.TrimLeftFunc
     0.01s  0.14%  2.17%      0.01s  0.14%  bufio.(*Reader).bufio.reset
     0.01s  0.14%  2.32%      0.01s  0.14%  bufio.(*Writer).WriteString
     0.01s  0.14%  2.46%      0.01s  0.14%  bytes.(*Buffer).grow
     0.01s  0.14%  2.61%      0.01s  0.14%  bytes.(*Buffer).tryGrowByReslice
     0.01s  0.14%  2.75%      0.01s  0.14%  encoding/json.(*arrayEncoder).encode
```

After:

```
Showing top 10 nodes out of 153
      flat  flat%   sum%        cum   cum%
     0.02s   0.3%   0.3%      0.02s   0.3%  bytes.(*Buffer).WriteString
     0.02s   0.3%  0.61%      0.02s   0.3%  context.(*cancelCtx).cancel
     0.02s   0.3%  0.91%      0.03s  0.46%  context.(*valueCtx).Value
     0.02s   0.3%  1.22%      0.02s   0.3%  github.com/TykTechnologies/tyk/config.Global (inline)
     0.02s   0.3%  1.52%      0.20s  3.04%  main.createMiddleware.func1.1
     0.02s   0.3%  1.82%      0.02s   0.3%  net/textproto.CanonicalMIMEHeaderKey
     0.02s   0.3%  2.13%      0.02s   0.3%  sync.(*RWMutex).RLock
     0.01s  0.15%  2.28%      0.01s  0.15%  bytes.(*Buffer).Write
     0.01s  0.15%  2.43%      0.01s  0.15%  bytes.(*Buffer).tryGrowByReslice (inline)
     0.01s  0.15%  2.58%      0.01s  0.15%  context.(*emptyCtx).Value
```