@buger 2018-05-08
Do you think it makes sense instead of comparing profiling information, actually compare benchmarks output?

And pls try using https://godoc.org/golang.org/x/tools/cmd/benchcmp for this