@dencoded 2018-05-08
Thats a pretty minor change but it could contribute to other code performance improvement, where this function is used. However if you compare benchmarks for some MW this improvement might be "diluted" as this function could be a small fraction of lots of logic running under the MW's hood.

So, for that kind of changes I would create dedicated benchmark to do more granular performance analysis - i.e. create bench mark for `func (s *SessionState) Hash() string` to prove that this piece of code is faster now (past `benchcmp` output) as Leon suggested. This function is actually doing two heavy things - user session state encoding and jungling with string so I think it is worth to have bench mark for it anyways (say we switch from message pack to something else and it becomes x5 time faster!)